BUG: All tests weren't working in integration tests

+++ tests/expr7.py
@@ -1,16 +1,16 @@
 def test_pow():
-    a: i32
+    a: f64
     a = pow(2, 2)
 
 
-def test_pow_1(a: i32, b: i32) -> i32:
-    res: i32
+def test_pow_1(a: i32, b: i32) -> f64:
+    res: f64
     res = pow(a, b)
     return res
 
 def main0():
     test_pow()
-    c: i32
+    c: f64
     c = test_pow_1(1, 2)
 
 main0()
